requirements.txt
django==1.10
django-crispy-forms==1.6.1
djangorestframework==3.6.3
django-pagedown==0.1.1
